fixed import problem so tests should pass
