+++ helixio/data_structures.py
@@ -1,0 +1,5 @@
+class AgentTelemetry:
+    arm_status = False
+    geodetic = [0, 0, 0]
+    position_ned = [0, 0, 0]
+    velocity_ned = [0, 0, 0]
